Move onceOnly in IOPlatform
