+++ core/jvm/src/main/scala/cats/effect/internals/IOPlatform.scala
@@ -16,6 +16,7 @@
 
 package cats.effect.internals
 
+import java.util.concurrent.atomic.AtomicBoolean
 import java.util.concurrent.locks.AbstractQueuedSynchronizer
 
 import cats.effect.IO
@@ -61,6 +62,17 @@ private[effect] object IOPlatform {
       case Right(a) => Some(a)
       case Left(ex) => throw ex
     }
+  }
+
+  /**
+   * Given any side-effecting function, builds a new one
+   * that has the idempotency property, making sure that its
+   * side-effects get triggered only once
+   */
+  def onceOnly[A](f: A => Unit): A => Unit = {
+    val wasCalled = new AtomicBoolean(false)
+
+    a => if (wasCalled.getAndSet(true)) () else f(a)
   }
 
   private final class OneShotLatch extends AbstractQueuedSynchronizer {
