--- core/jvm/src/main/scala/cats/effect/internals/Utils.scala
@@ -1,32 +1,0 @@
-/*
- * Copyright 2017 Typelevel
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cats.effect.internals
-
-import java.util.concurrent.atomic.AtomicBoolean
-
-private[effect] object Utils {
-  /**
-   * Given any side-effecting function, builds a new one
-   * that has the idempotency property, making sure that its
-   * side-effects get triggered only once
-   */
-  def onceOnly[A](f: A => Unit): A => Unit = {
-    val wasCalled = new AtomicBoolean(false)
-
-    a => if (wasCalled.getAndSet(true)) () else f(a)
-  }
-}
